restructuring; use mocha for js unit tests
