+++ package.json
@@ -11,13 +11,16 @@
     "test"
   ],
   "scripts": {
-    "test": "scripts/test.sh",
+    "test": "npm run test:contracts && npm run test:utils",
+    "test:contracts": "scripts/test.sh test/contracts/*",
+    "test:utils": "./node_modules/.bin/mocha --require @babel/register --require @babel/polyfill test/utils/",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "lint:sol": "solhint contracts/*.{sol,slb}",
     "lint:all": "npm run lint && npm run lint:sol",
     "lint:all:fix": "npm run lint:fix && npm run lint:sol",
-    "console": "truffle console",
+    "console": "./node_modules/.bin/truffle console",
+    "compile:contracts": "./node_modules/.bin/truffle compile",
     "version": "scripts/version.js",
     "coverage": "scripts/coverage.sh"
   },
