+++ test/helpers/utils.js
@@ -1,4 +1,4 @@
-const OP = require('./../utils/constants');
+const OP = require('./../../utils/constants');
 const ethers = require('ethers');
 const { PUSH1 } = OP;
 
@@ -167,7 +167,8 @@ export const getCode = (fixture) => {
 };
 export const wallets = [];
 
-export const provider = new ethers.providers.Web3Provider(web3.currentProvider);
+export const provider =
+  typeof web3 !== 'undefined' ? new ethers.providers.Web3Provider(web3.currentProvider) : undefined;
 
 for (var i = 0; i < 10; i++) {
   const privateKey = '0x2bdd21761a483f71054e14f5b827213567971c676928d9a1808cbfa4b750120' + i;
